Add files via upload
sales data dashboard which has been created on power bi and power query illustartes aspects of products and how many times these products sale over the world? and orders and we calculate some of measures to get out sum of taxAmt ,sum of frieght , sum of unitprice , sum of orders and count of territory group.